Linked Users to their personal to do lists

+++ convex/functions.ts
@@ -1,10 +1,15 @@
 import { mutation, query } from "./_generated/server";
 import { v } from "convex/values"
+import { requireUser } from "./helpers";
 
 
 export const listTodos = query({
     handler: async (ctx) => {
-        return await ctx.db.query("todos").collect();
+        const user = await requireUser(ctx);
+        return await ctx.db.query("todos")
+            .withIndex("by_user_id", q => q.eq("userID", user.tokenIdentifier))
+            .collect();
+        
     }
 });
 
@@ -14,10 +19,12 @@ export const createTodo = mutation({
         description: v.string(),
     },
     handler: async (ctx, args) => {
+        const user = await requireUser(ctx);
         await ctx.db.insert("todos", {
             title: args.title,
             description: args.description,
-            completed: false
+            completed: false,
+            userID: user.tokenIdentifier
         })
     }, 
 });
@@ -28,6 +35,11 @@ export const updateTodo = mutation({
         completed: v.boolean(),
     },
     handler: async (ctx, args) => {
+        const user = await requireUser(ctx);
+        const todo = await ctx.db.get(args.id);
+        if (todo?.userID !== user.tokenIdentifier) {
+            throw new Error("Unauthorized");
+        }
         await ctx.db.patch(args.id, {
             completed: args.completed
         })
@@ -39,6 +51,11 @@ export const deleteTodo = mutation({
         id: v.id("todos")
     },
     handler: async (ctx, args) => {
+        const user = await requireUser(ctx);
+        const todo = await ctx.db.get(args.id);
+        if (todo?.userID !== user.tokenIdentifier) {
+            throw new Error("Unauthorized");
+        }
         await ctx.db.delete(args.id)
     }
 });
